Refactor NWB modules with harmonization layer
- Rename nwb.py -> nwb_predictions.py for clarity
- Rename nwb_ann.py -> nwb_annotations.py for consistency
- Create new nwb.py harmonization layer with:
  - NwbFormat enum for format selection
  - load_nwb() that auto-detects PoseTraining vs PoseEstimation
  - save_nwb() that auto-detects format based on instance types
- Remove append functionality from high-level API for simplicity
- Update all imports and tests accordingly

This provides clearer separation between prediction and annotation
formats while maintaining backwards compatibility through the
harmonization layer.
