+++ tests/io/test_main.py
@@ -25,21 +25,24 @@ def test_load_slp(slp_typical):
     assert type(load_file(slp_typical)) is Labels
 
 
-def test_nwb(tmp_path, slp_typical):
-    labels = load_slp(slp_typical)
-    save_nwb(labels, tmp_path / "test_nwb.nwb")
-    loaded_labels = load_nwb(tmp_path / "test_nwb.nwb")
+def test_nwb(tmp_path, slp_typical, slp_real_data):
+    # Test with predictions (slp_typical has predictions)
+    labels_pred = load_slp(slp_typical)
+    save_nwb(labels_pred, tmp_path / "test_nwb_pred.nwb")
+    loaded_labels = load_nwb(tmp_path / "test_nwb_pred.nwb")
     assert type(loaded_labels) is Labels
-    assert type(load_file(tmp_path / "test_nwb.nwb")) is Labels
-    assert len(loaded_labels) == len(labels)
-
-    labels2 = load_slp(slp_typical)
-    labels2.videos[0].filename = "test"
-    save_nwb(labels2, tmp_path / "test_nwb.nwb", append=True)
-    loaded_labels = load_nwb(tmp_path / "test_nwb.nwb")
+    assert type(load_file(tmp_path / "test_nwb_pred.nwb")) is Labels
+    
+    # Test with annotations (slp_real_data has user instances)
+    labels_ann = load_slp(slp_real_data)
+    save_nwb(labels_ann, tmp_path / "test_nwb_ann.nwb")
+    loaded_labels = load_nwb(tmp_path / "test_nwb_ann.nwb")
     assert type(loaded_labels) is Labels
-    assert len(loaded_labels) == (len(labels) + len(labels2))
-    assert len(loaded_labels.videos) == 2
+    
+    # Test overwriting (no append)
+    save_nwb(labels_pred, tmp_path / "test_nwb_pred.nwb")  # Overwrites
+    loaded_labels = load_nwb(tmp_path / "test_nwb_pred.nwb")
+    assert type(loaded_labels) is Labels
 
 
 def test_labelstudio(tmp_path, slp_typical):
